# More test optimisations