--- a/ocean_lib/data_provider/data_service_provider.py
+++ b/ocean_lib/data_provider/data_service_provider.py
@@ -10,14 +10,12 @@ from json import JSONDecodeError
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Union
 
-import requests
 from enforce_typing import enforce_types
 from requests.models import PreparedRequest, Response
 
 from ocean_lib.agreements.service_types import ServiceTypes
 from ocean_lib.data_provider.base import DataServiceProviderBase
 from ocean_lib.data_provider.fileinfo_provider import FileInfoProvider
-from ocean_lib.exceptions import DataProviderException
 from ocean_lib.http_requests.requests_session import get_requests_session
 from ocean_lib.models.compute_input import ComputeInput
 from ocean_lib.structures.algorithm_metadata import AlgorithmMetadata
@@ -45,7 +43,7 @@ class DataServiceProvider(DataServiceProviderBase):
         userdata: Optional[Dict] = None,
     ) -> Response:
 
-        _, initialize_endpoint = DataServiceProvider.build_endpoint(
+        method, initialize_endpoint = DataServiceProvider.build_endpoint(
             "initialize", service.service_endpoint
         )
 
@@ -60,20 +58,12 @@ class DataServiceProvider(DataServiceProviderBase):
             payload["userdata"] = userdata
 
         response = DataServiceProvider._http_method(
-            "get", url=initialize_endpoint, params=payload
+            method, url=initialize_endpoint, params=payload
         )
-        if not response or not hasattr(response, "status_code"):
-            raise DataProviderException(
-                f"Failed to get a response for request: initializeEndpoint={initialize_endpoint}, payload={payload}, response is {response}"
-            )
 
-        if response.status_code != 200:
-            msg = (
-                f"Initialize service failed at the initializeEndpoint "
-                f"{initialize_endpoint}, reason {response.text}, status {response.status_code}"
-            )
-            logger.error(msg)
-            raise DataProviderException(msg)
+        DataServiceProviderBase.check_response(
+            response, "initializeEndpoint", initialize_endpoint, payload
+        )
 
         logger.info(
             f"Service initialized successfully"
@@ -98,7 +88,7 @@ class DataServiceProvider(DataServiceProviderBase):
         The first dataset is also required to have a compute service.
         """
         (
-            _,
+            method,
             initialize_compute_endpoint,
         ) = DataServiceProvider.build_endpoint("initializeCompute", service_endpoint)
 
@@ -113,7 +103,7 @@ class DataServiceProvider(DataServiceProviderBase):
         }
 
         response = DataServiceProvider._http_method(
-            "post",
+            method,
             initialize_compute_endpoint,
             data=json.dumps(payload),
             headers={"content-type": "application/json"},
@@ -154,7 +144,7 @@ class DataServiceProvider(DataServiceProviderBase):
             )
             indexes = [index]
 
-        _, download_endpoint = DataServiceProvider.build_endpoint(
+        method, download_endpoint = DataServiceProvider.build_endpoint(
             "download", service_endpoint
         )
 
@@ -175,7 +165,7 @@ class DataServiceProvider(DataServiceProviderBase):
                 consumer_wallet, did
             )
             response = DataServiceProvider._http_method(
-                "get", url=download_endpoint, params=payload, stream=True, timeout=3
+                method, url=download_endpoint, params=payload, stream=True, timeout=3
             )
 
             DataServiceProviderBase.check_response(
@@ -237,8 +227,7 @@ class DataServiceProvider(DataServiceProviderBase):
         _, compute_endpoint = DataServiceProvider.build_endpoint(
             "computeStart", dataset_compute_service.service_endpoint
         )
-        response = DataServiceProvider._http_method(
-            "post",
+        response = DataServiceProvider.post_raw(
             compute_endpoint,
             data=json.dumps(payload),
             headers={"content-type": "application/json"},
@@ -363,7 +352,7 @@ class DataServiceProvider(DataServiceProviderBase):
             "consumerAddress": consumer.address,
         }
 
-        (_, compute_job_result_endpoint,) = DataServiceProvider.build_endpoint(
+        (_, compute_job_result_endpoint) = DataServiceProvider.build_endpoint(
             "computeResult", dataset_compute_service.service_endpoint
         )
         req.prepare_url(compute_job_result_endpoint, params)
@@ -444,8 +433,10 @@ class DataServiceProvider(DataServiceProviderBase):
         )
 
         resp_content = json.loads(response.content.decode("utf-8"))
+
         if isinstance(resp_content, list):
             return resp_content[0]
+
         return resp_content
 
     @staticmethod
@@ -469,18 +460,13 @@ class DataServiceProvider(DataServiceProviderBase):
                 f"for `algorithm_meta`, got {type(algorithm_meta)}"
             )
 
+        input_datasets = input_datasets if input_datasets else []
         _input_datasets = []
-        if input_datasets:
-            for _input in input_datasets:
-                assert _input.did, "The received dataset does not have a did."
-                assert (
-                    _input.transfer_tx_id
-                ), "The received dataset does not have a transaction id."
-                assert (
-                    _input.service_id
-                ), "The received dataset does not have a specified service id."
-                if _input.did != dataset.did:
-                    _input_datasets.append(_input.as_dictionary())
+        for _input in input_datasets:
+            for req_key in ["did", "transfer_tx_id", "service_id"]:
+                assert getattr(
+                    _input, req_key
+                ), f"The received dataset does not have a {req_key}."
 
         nonce, signature = DataServiceProvider.sign_message(
             consumer, f"{consumer.address}{dataset.did}"
@@ -547,13 +533,9 @@ class DataServiceProvider(DataServiceProviderBase):
         if userdata is not None:
             data["userdata"] = userdata
 
-        response = requests.post(endpoint, json=data)
+        response = DataServiceProvider.post_raw(endpoint, json=data)
 
         if not response or response.status_code != 200:
             return False
 
-        response = response.json()
-        for ddo_info in response:
-            return ddo_info["valid"]
-
-        return False
+        return any([file_info["valid"] for file_info in response.json()])