--- a/ocean_lib/web3_internal/utils.py
+++ b/ocean_lib/web3_internal/utils.py
@@ -4,7 +4,7 @@
 #
 import logging
 from collections import namedtuple
-from typing import Any
+from typing import Any, Union
 
 import requests
 from brownie import network
@@ -32,7 +32,13 @@ def to_32byte_hex(val: int) -> str:
 
 
 @enforce_types
-def sign_with_key(message_hash: HexBytes, key: str) -> str:
+def sign_with_key(message_hash: Union[HexBytes, str], key: str) -> str:
+    if isinstance(message_hash, str):
+        message_hash = Web3.solidityKeccak(
+            ["bytes"],
+            [Web3.toBytes(text=message_hash)],
+        )
+
     pk = keys.PrivateKey(Web3.toBytes(hexstr=key))
     prefix = "\x19Ethereum Signed Message:\n32"
     signable_hash = Web3.solidityKeccak(