More work on #929. I also did some cleanup of the provider functions.
Initially, I wanted to use the methods from build_service_endpoint (first item of the tuple returned). However, it seems it's not entirely accurate and so I kept the hardcoding. I would suggest removing the method part entirely, but we can discuss that later.

In regards to tests, here are the only tests that now take > 10s:
```
93.92s call     tests/integration/ganache/test_compute_flow.py::TestComputeFlow::test_compute_update_trusted_algorithm
54.45s call     tests/readmes/test_readmes.py::TestReadmes::test_script_execution[test_c2d-flow.py]
35.07s call     tests/integration/ganache/test_compute_flow.py::TestComputeFlow::test_compute_reuse_order
26.85s call     tests/readmes/test_readmes.py::TestReadmes::test_script_execution[test_main-flow.py]
24.74s call     ocean_lib/ocean/test/test_ocean_assets.py::test_create_pricing_schemas
21.90s call     ocean_lib/ocean/test/test_ocean_assets.py::test_update_datatokens
19.99s call     tests/readmes/test_readmes.py::TestReadmes::test_script_execution[test_search-and-filter-assets.py]
17.20s call     tests/readmes/test_readmes.py::TestReadmes::test_script_execution[test_custody-light-flow.py]
15.02s call     ocean_lib/ocean/test/test_ocean_assets.py::test_update
11.00s call     ocean_lib/models/ve/test/test_ve_ocean.py::test_ve_ocean1
10.27s call     tests/integration/ganache/test_graphql.py::test_consume_parametrized_graphql_query
```

Mostly, large tests are large because they handle updates and sequential work that needs to wait for Aquarius. There is no way of restructuring those tests without losing meaning. For readme tests, there is no way I can do async publishing because it would complicate the readmes. 

For compute tests, the improvement is 186.21s to 105.35s.